requirements.txt
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
six==1.16.0

Pillow~=9.4.0